requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
langchain>=0.1.0
langchain-community>=0.0.10
openai==1.3.5
faiss-cpu==1.7.4
python-multipart==0.0.6
beautifulsoup4==4.12.2
pydantic==2.5.2
requests==2.31.0
python-dotenv==1.0.0